Restructure packages to the adapters and internal
