+++ sandwich/src/main/kotlin/com/skydoves/sandwich/adapters/ApiResponseCallAdapterFactory.kt
@@ -14,8 +14,12 @@
  * limitations under the License.
  */
 
-package com.skydoves.sandwich
+@file:Suppress("unused")
 
+package com.skydoves.sandwich.adapters
+
+import com.skydoves.sandwich.ApiResponse
+import retrofit2.Call
 import retrofit2.CallAdapter
 import retrofit2.Retrofit
 import java.lang.reflect.ParameterizedType
@@ -24,33 +28,38 @@ import java.lang.reflect.Type
 /**
  * @author skydoves (Jaewoong Eum)
  *
- * DataSourceCallAdapterFactory is an call adapter factory for creating [DataSource].
+ * CoroutinesResponseCallAdapterFactory is an coroutines call adapter factory for creating [ApiResponse].
  *
- * Adding this class to [Retrofit] allows you to return on [DataSource] from service method.
+ * Adding this class to [Retrofit] allows you to return on [ApiResponse] from service method.
  *
  * ```
  * @GET("DisneyPosters.json")
- * fun fetchDisneyPosterList(): DataSource<List<Poster>>
+ * suspend fun fetchDisneyPosterList(): ApiResponse<List<Poster>>
  * ```
  */
-public class DataSourceCallAdapterFactory private constructor() : CallAdapter.Factory() {
+public class ApiResponseCallAdapterFactory private constructor() : CallAdapter.Factory() {
 
   override fun get(
     returnType: Type,
     annotations: Array<Annotation>,
     retrofit: Retrofit
-  ): CallAdapter<*, *>? {
-    val rawType = getRawType(returnType)
-    if (rawType != DataSource::class.java) {
-      return null
+  ): ApiResponseCallAdapter? = when (getRawType(returnType)) {
+    Call::class.java -> {
+      val callType = getParameterUpperBound(0, returnType as ParameterizedType)
+      when (getRawType(callType)) {
+        ApiResponse::class.java -> {
+          val resultType = getParameterUpperBound(0, callType as ParameterizedType)
+          ApiResponseCallAdapter(resultType)
+        }
+        else -> null
+      }
     }
-    val enclosingType = returnType as ParameterizedType
-    val actualType = enclosingType.actualTypeArguments[0]
-    return DataSourceCallAdapter<Type>(actualType)
+    else -> null
   }
 
   public companion object {
     @JvmStatic
-    public fun create(): DataSourceCallAdapterFactory = DataSourceCallAdapterFactory()
+    public fun create(): ApiResponseCallAdapterFactory =
+      ApiResponseCallAdapterFactory()
   }
 }
