+++ sandwich/src/main/kotlin/com/skydoves/sandwich/coroutines/CoroutinesDataSourceCallAdapter.kt
@@ -17,6 +17,7 @@
 package com.skydoves.sandwich.coroutines
 
 import com.skydoves.sandwich.DataSource
+import com.skydoves.sandwich.adapters.internal.DataSourceCallDelegate
 import retrofit2.Call
 import retrofit2.CallAdapter
 import java.lang.reflect.Type
